Rendering cursor and highlighting in the graphics driver. Added 15-bit color mode support. Dithering in 15-bit and 16-bit color mode.

+++ Libraries/perception/public/perception/draw.h
@@ -18,25 +18,28 @@
 
 namespace perception {
 
-void DrawSprite1bitAlpha(size_t x, size_t y, uint32 *sprite, size_t width, size_t height,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height, size_t minx, size_t miny, size_t maxx, size_t maxy);
-void DrawSpriteAlpha(size_t x, size_t y, uint32 *sprite, size_t width, size_t height,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height, size_t minx, size_t miny, size_t maxx, size_t maxy);
-void DrawSprite(size_t x, size_t y, uint32 *sprite, size_t width, size_t height,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height, size_t minx, size_t miny, size_t maxx, size_t maxy);
-void DrawXLine(size_t x, size_t y, size_t width, uint32 colour,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height);
-void DrawXLineAlpha(size_t x, size_t y, size_t width, uint32 colour,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height);
-void DrawYLine(size_t x, size_t y, size_t height, uint32 colour,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height);
-void DrawYLineAlpha(size_t x, size_t y, size_t height, uint32 colour,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height);
-void PlotPixel(size_t x, size_t y, uint32 colour,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height);
-void FillRectangle(size_t minx, size_t miny, size_t maxx, size_t maxy, uint32 colour,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height);
-void FillRectangleAlpha(size_t minx, size_t miny, size_t maxx, size_t maxy, uint32 colour,
-	uint32 *buffer, size_t buffer_width, size_t buffer_height);
+void DrawSprite1bitAlpha(int x, int y, uint32 *sprite, int width, int height,
+	uint32 *buffer, int buffer_width, int buffer_height, int minx, int miny,
+	int maxx, int maxy);
+void DrawSpriteAlpha(int x, int y, uint32 *sprite, int width,
+	int height, uint32 *buffer, int buffer_width, int buffer_height,
+	int minx, int miny, int maxx, int maxy);
+void DrawSprite(int x, int y, uint32 *sprite, int width, int height,
+	uint32 *buffer, int buffer_width, int buffer_height, int minx,
+	int miny, int maxx, int maxy);
+void DrawXLine(int x, int y, int width, uint32 colour,
+	uint32 *buffer, int buffer_width, int buffer_height);
+void DrawXLineAlpha(int x, int y, int width, uint32 colour,
+	uint32 *buffer, int buffer_width, int buffer_height);
+void DrawYLine(int x, int y, int height, uint32 colour,
+	uint32 *buffer, int buffer_width, int buffer_height);
+void DrawYLineAlpha(int x, int y, int height, uint32 colour,
+	uint32 *buffer, int buffer_width, int buffer_height);
+void PlotPixel(int x, int y, uint32 colour,
+	uint32 *buffer, int buffer_width, int buffer_height);
+void FillRectangle(int minx, int miny, int maxx, int maxy,
+	uint32 colour, uint32 *buffer, int buffer_width, int buffer_height);
+void FillRectangleAlpha(int minx, int miny, int maxx, int maxy,
+	uint32 colour, uint32 *buffer, int buffer_width, int buffer_height);
 	
 }
